Add more automations
