+++ cmd/go.mod
@@ -5,5 +5,6 @@ go 1.14
 require (
 	github.com/go-git/go-git/v5 v5.0.0
 	github.com/google/go-github/v28 v28.1.1
+	github.com/jszwedko/go-circleci v0.3.0
 	golang.org/x/oauth2 v0.0.0-20200107190931-bf48bf16ab8d
 )
